Fix rubocop Style/MethodName "Use snake_case for method names."
Rename private method
Add deprecation warnings for two public methods
Delete unused method
Update .rubocop.yml accordingly
Refactor duplicated chassis power command methods
