+++ lib/rubyipmi/freeipmi/commands/power.rb
@@ -1,5 +1,9 @@
+require 'rubyipmi/commands/mixins/power_mixin'
+
 module Rubyipmi::Freeipmi
   class Power < Rubyipmi::Freeipmi::BaseCommand
+    include Rubyipmi::PowerMixin
+
     def initialize(opts = ObservableHash.new)
       super("ipmipower", opts)
     end
@@ -12,37 +16,7 @@ module Rubyipmi::Freeipmi
       @result
     end
 
-    # Turn on the system
-    def on
-      command("on")
-    end
-
-    # Turn off the system
-    def off
-      command("off")
-    end
-
-    # Power cycle the system
-    def cycle
-      # if the system is off turn it on
-      if off?
-        on
-      else
-        command("cycle")
-      end
-    end
-
-    # Perform a power reset on the system
-    def reset
-      command("reset")
-    end
-
-    # Perform a soft shutdown, like briefly pushing the power button
-    def softShutdown
-      command("soft")
-    end
-
-    def powerInterrupt
+    def power_interrupt
       command("pulse")
     end
 
@@ -50,16 +24,6 @@ module Rubyipmi::Freeipmi
     def status
       value = command("stat")
       @result.split(":").last.chomp.strip if value
-    end
-
-    # Test to see if the power is on
-    def on?
-      status == "on"
-    end
-
-    # Test to see if the power is off
-    def off?
-      status == "off"
     end
   end
 end
