+++ lib/rubyipmi/ipmitool/commands/power.rb
@@ -1,5 +1,9 @@
+require 'rubyipmi/commands/mixins/power_mixin'
+
 module Rubyipmi::Ipmitool
   class Power < Rubyipmi::Ipmitool::BaseCommand
+    include Rubyipmi::PowerMixin
+
     def initialize(opts = ObservableHash.new)
       super("ipmitool", opts)
     end
@@ -12,45 +16,7 @@ module Rubyipmi::Ipmitool
       value
     end
 
-    # Turn on the system
-    def on
-      if on?
-        return true
-      else
-        command("on")
-      end
-    end
-
-    # Turn off the system
-    def off
-      if off?
-        return true
-      else
-        command("off")
-      end
-    end
-
-    # Power cycle the system
-    def cycle
-      # if the system is off turn it on
-      if off?
-        on
-      else
-        command("cycle")
-      end
-    end
-
-    # Perform a power reset on the system
-    def reset
-      command("reset")
-    end
-
-    # Perform a soft shutdown, like briefly pushing the power button
-    def softShutdown
-      command("soft")
-    end
-
-    def powerInterrupt
+    def power_interrupt
       command("diag")
     end
 
@@ -58,16 +24,6 @@ module Rubyipmi::Ipmitool
     def status
       value = command("status")
       @result.match(/(off|on)/).to_s if value
-    end
-
-    # Test to see if the power is on
-    def on?
-      status == "on"
-    end
-
-    # Test to see if the power is off
-    def off?
-      status == "off"
     end
   end
 end
