+++ lib/rubyipmi/ipmitool/errorcodes.rb
@@ -28,28 +28,6 @@ module Rubyipmi
         end
         raise "No Fix found" if fix.nil?
       end
-
-      def throwError
-        # Find out what kind of error is happening, parse results
-        # Check for authentication or connection issue
-
-        if @result =~ /timeout|timed\ out/
-          code = "ipmi call: #{@lastcall} timed out"
-          raise code
-        else
-          code = @result.split(":").last.chomp.strip unless @result.empty?
-        end
-        case code
-        when "invalid hostname"
-          raise code
-        when "password invalid"
-          raise code
-        when "username invalid"
-          raise code
-        else
-          raise :ipmierror, code
-        end
-      end
     end
   end
 end
